requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudinary==1.28.0
Flask==2.0.2
Flask-Assets==2.0
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mysql-client==0.0.1
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
pandas==1.3.5
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.3.3
tensorflow-estimator==2.3.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.7
webassets==2.0
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
